chore: stop blocking querries
When I set the file descriptor to non-blocking, the error was EBUSY.
When using filters, I could sometimes wait for as long as 13s, an
unacceptable delay. I have removed all use of kernel filters (i.e. I
think that's what the lib uses), but I don't know how this filtering
mechanism works and why I get an EBUSY

+++ include/database.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 void db_setup(void);
-void db_save(uint64_t timestamp, uint64_t duration_ms, float rx_mb, float tx_mb);
+void db_save(uint64_t timestamp, uint32_t duration_ms, float rx_mb, float tx_mb);
 
 
 #ifdef __cplusplus
